# Faster optimized frozen dictionary creation (5/n) 